+++ src/Blocks/index.tsx
@@ -1,9 +1,9 @@
 import { ElementType } from 'react';
 
 import { Block } from '@src/Block';
-import { FlexGridProps } from '@src/types';
+import { MerengueData } from '@src/types';
 
-export const Blocks: ElementType<Required<Pick<FlexGridProps, 'blocks'>>> = ({ blocks }) => (
+export const Blocks: ElementType<Required<Pick<MerengueData, 'blocks'>>> = ({ blocks }) => (
   <>
     {blocks.map(({ renderContent, ...block }, blockKey) => (
       <Block key={blockKey} {...block}>
