+++ src/stories/MerengueBox.stories.tsx
@@ -1,12 +1,13 @@
-import { Story } from '@ladle/react';
 import React, { useRef } from 'react';
 
+import { Story } from '@ladle/react';
+
 import { Block } from '../Block';
-import { FlexGrid } from '../FlexGrid';
-import { FlexGridProps } from '../types';
+import { MerengueBox } from '../MerengueBox';
+import { MerengueData } from '../types';
 import { data } from './utils/data';
 
-export const FromDataObject: Story<FlexGridProps> = props => <FlexGrid {...data} {...props} />;
+export const FromDataObject: Story<MerengueData> = props => <MerengueBox {...data} {...props} />;
 
 FromDataObject.args = {
   extendContent: false,
@@ -16,11 +17,11 @@ FromDataObject.args = {
   as: 'div',
 };
 
-export const Nested: Story<FlexGridProps> = props => {
+export const Nested: Story<MerengueData> = props => {
   const ref = useRef(null);
   const ref2 = useRef(null);
   return (
-    <FlexGrid {...props} as="ol" ref={ref}>
+    <MerengueBox {...props} as="ol" ref={ref}>
       <Block as="li" ref={ref2} onClick={() => console.log({ ref, ref2 })}>
         Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellat optio nemo, animi eligendi
         voluptatem voluptatum. Sit facere fugit laudantium adipisci itaque similique incidunt, quos
@@ -51,7 +52,7 @@ export const Nested: Story<FlexGridProps> = props => {
         voluptatem voluptatum. Sit facere fugit laudantium adipisci itaque similique incidunt, quos
         assumenda totam! Eaque magni culpa quisquam.
       </Block>
-    </FlexGrid>
+    </MerengueBox>
   );
 };
